refactor clustering files
