+++ aeon/clustering/__init__.py
@@ -8,8 +8,8 @@ __all__ = [
 ]
 __author__ = ["chrisholder", "TonyBagnall"]
 
+from aeon.clustering._clara import TimeSeriesCLARA
+from aeon.clustering._clarans import TimeSeriesCLARANS
+from aeon.clustering._k_means import TimeSeriesKMeans
+from aeon.clustering._k_medoids import TimeSeriesKMedoids
 from aeon.clustering.base import BaseClusterer
-from aeon.clustering.clara import TimeSeriesCLARA
-from aeon.clustering.clarans import TimeSeriesCLARANS
-from aeon.clustering.k_means import TimeSeriesKMeans
-from aeon.clustering.k_medoids import TimeSeriesKMedoids
