+++ aeon/clustering/_clarans.py
@@ -1,25 +1,27 @@
 """Time series kmedoids."""
 __author__ = ["chrisholder", "TonyBagnall"]
 
+import math
 from typing import Callable, Union
 
 import numpy as np
 from numpy.random import RandomState
-from sklearn.utils import check_random_state
 
-from aeon.clustering.base import BaseClusterer
-from aeon.clustering.k_medoids import TimeSeriesKMedoids
+from aeon.clustering._k_medoids import TimeSeriesKMedoids
 
 
-class TimeSeriesCLARA(BaseClusterer):
-    """Time series CLARA implementation.
+class TimeSeriesCLARANS(TimeSeriesKMedoids):
+    """Time series CLARANS implementation.
 
-    Clustering LARge Applications (CLARA) [1] is a clustering algorithm that
-    samples the dataset, applies PAM to the sample, and then uses the
-    medoids from the sample to seed PAM on the entire dataset.
+    CLARA based raNdomised Search (CLARANS) [1] adapts the swap operation of PAM to
+    use a more greedy approach. This is done by only performing the first swap which
+    results in a reduction in total deviation before continuing evaluation. It limits
+    the number of attempts known as max neighbours to randomly select and check if
+    total deviation is reduced. This random selection gives CLARANS an advantage when
+    handling large datasets by avoiding local minima.
 
-    For a comparison of using CLARA for time series compared to other k-medoids
-    algorithms see [3].
+    For a comparison of using CLARANS for time series compared to other k-medoids
+    algorithms see [2].
 
     Parameters
     ----------
@@ -42,23 +44,16 @@ class TimeSeriesCLARA(BaseClusterer):
         strings for metrics can be found in the documentation for
         :func:`aeon.distances.get_distance_function`. If a callable is passed it must be
         a function that takes two 2d numpy arrays as input and returns a float.
-    n_samples : int, default=None,
-        Number of samples to sample from the dataset. If None, then
-        min(n_instances, 40 + 2 * n_clusters) is used.
-    n_sampling_iters : int, default=5,
-        Number of different subsets of samples to try. The best subset cluster centers
-        are used.
+    max_neighbours : int, default=None,
+        The maximum number of neighbouring solutions that the algorithm will explore
+        for each set of medoids. A neighbouring solution is obtained by replacing
+        one of the medoids with a non-medoid and seeing if total cost reduces. If
+        not specified max_neighbours is set to 1.25% of the total number of possible
+        swaps (as suggested in the orginal paper).
     n_init : int, default=5
         Number of times the PAM algorithm will be run with different
         centroid seeds. The final result will be the best output of n_init
         consecutive runs in terms of inertia.
-    max_iter : int, default=300
-        Maximum number of iterations of the PAM algorithm for a single
-        run.
-    tol : float, default=1e-6
-        Relative tolerance with regards to Frobenius norm of the difference
-        in the cluster centers of two consecutive iterations to declare
-        convergence.
     verbose : bool, default=False
         Verbosity mode.
     random_state : int or np.random.RandomState instance or None, default=None
@@ -80,21 +75,22 @@ class TimeSeriesCLARA(BaseClusterer):
 
     Examples
     --------
-    >>> from aeon.clustering import TimeSeriesCLARA
+    >>> from aeon.clustering import TimeSeriesCLARANS
     >>> from aeon.datasets import load_basic_motions
     >>> # Load data
     >>> X_train, y_train = load_basic_motions(split="TRAIN")[0:10]
     >>> X_test, y_test = load_basic_motions(split="TEST")[0:10]
     >>> # Example of PAM clustering
-    >>> km = TimeSeriesCLARA(n_clusters=3, distance="euclidean", random_state=1)
+    >>> km = TimeSeriesCLARANS(n_clusters=3, distance="euclidean", random_state=1)
     >>> km.fit(X_train)
-    TimeSeriesCLARA(distance='euclidean', n_clusters=3, random_state=1)
+    TimeSeriesCLARANS(distance='euclidean', n_clusters=3, random_state=1)
     >>> preds = km.predict(X_test)
 
     References
     ----------
-    .. [1] Kaufman, Leonard & Rousseeuw, Peter. (1986). Clustering Large Data Sets.
-    10.1016/B978-0-444-87877-9.50039-X.
+    .. [1] R. T. Ng and Jiawei Han, "CLARANS: a method for clustering objects for
+    spatial data mining," in IEEE Transactions on Knowledge and Data Engineering,
+    vol. 14, no. 5, pp. 1003-1016, Sept.-Oct. 2002, doi: 10.1109/TKDE.2002.1033770.
 
     .. [2] Holder, Christopher & Guijo-Rubio, David & Bagnall, Anthony. (2023).
     Clustering time series with k-medoids based algorithms.
@@ -102,92 +98,88 @@ class TimeSeriesCLARA(BaseClusterer):
     Data (AALTD 2023).
     """
 
-    _tags = {
-        "capability:multivariate": True,
-    }
-
     def __init__(
         self,
         n_clusters: int = 8,
         init_algorithm: Union[str, np.ndarray] = "random",
         distance: Union[str, Callable] = "msm",
-        n_samples: int = None,
-        n_sampling_iters: int = 10,
-        n_init: int = 1,
-        max_iter: int = 300,
-        tol: float = 1e-6,
+        max_neighbours: int = None,
+        n_init: int = 10,
         verbose: bool = False,
         random_state: Union[int, RandomState] = None,
         distance_params: dict = None,
     ):
-        self.init_algorithm = init_algorithm
-        self.distance = distance
-        self.n_init = n_init
-        self.max_iter = max_iter
-        self.tol = tol
-        self.verbose = verbose
-        self.random_state = random_state
-        self.distance_params = distance_params
-        self.n_samples = n_samples
-        self.n_sampling_iters = n_sampling_iters
+        self.max_neighbours = max_neighbours
 
-        self.cluster_centers_ = None
-        self.labels_ = None
-        self.inertia_ = None
-        self.n_iter_ = 0
+        super(TimeSeriesCLARANS, self).__init__(
+            n_clusters=n_clusters,
+            init_algorithm=init_algorithm,
+            distance=distance,
+            n_init=n_init,
+            verbose=verbose,
+            random_state=random_state,
+            distance_params=distance_params,
+        )
 
-        self._random_state = None
-        self._kmedoids_instance = None
+    def _fit_one_init(self, X: np.ndarray, max_neighbours: int):
+        j = 0
+        X_indexes = np.arange(X.shape[0], dtype=int)
+        if isinstance(self._init_algorithm, Callable):
+            best_medoids = self._init_algorithm(X)
+        else:
+            best_medoids = self._init_algorithm
+        best_non_medoids = np.setdiff1d(X_indexes, best_medoids)
+        best_cost = (
+            self._compute_pairwise(X, best_non_medoids, best_medoids).min(axis=1).sum()
+        )
+        num_non_medoids = X.shape[0] - self.n_clusters
+        while j < max_neighbours:
+            new_medoids = best_medoids.copy()
+            new_non_medoids = best_non_medoids.copy()
+            to_replace_index = self._random_state.randint(self.n_clusters)
+            replace_with_index = self._random_state.randint(num_non_medoids)
+            temp = new_medoids[to_replace_index]
+            new_medoids[to_replace_index] = new_non_medoids[replace_with_index]
+            new_non_medoids[replace_with_index] = temp
 
-        super(TimeSeriesCLARA, self).__init__(n_clusters)
+            new_cost = (
+                self._compute_pairwise(X, new_non_medoids, new_medoids)
+                .min(axis=1)
+                .sum()
+            )
+            if new_cost < best_cost:
+                best_cost = new_cost
+                best_medoids = new_medoids
+                best_non_medoids = np.setdiff1d(X_indexes, best_medoids)
+            else:
+                j += 1
 
-    def _predict(self, X: np.ndarray, y=None) -> np.ndarray:
-        return self._kmedoids_instance.predict(X)
+        return best_medoids, best_cost
 
     def _fit(self, X: np.ndarray, y=None):
-        self._random_state = check_random_state(self.random_state)
-        n_instances = X.shape[0]
-        if self.n_samples is None:
-            n_samples = max(
-                min(n_instances, 40 + 2 * self.n_clusters), self.n_clusters + 1
+        self._check_params(X)
+
+        max_neighbours = self.max_neighbours
+        if self.max_neighbours is None:
+            max_neighbours = math.ceil(
+                (1.25 / 100) * (self.n_clusters * (X.shape[0] - self.n_clusters))
             )
-        else:
-            n_samples = self.n_samples
 
-        best_score = np.inf
-        best_pam = None
-        for _ in range(self.n_sampling_iters):
-            sample_idxs = np.arange(n_samples)
-            if n_samples < n_instances:
-                sample_idxs = self._random_state.choice(
-                    sample_idxs,
-                    size=n_samples,
-                    replace=False,
-                )
-            pam = TimeSeriesKMedoids(
-                n_clusters=self.n_clusters,
-                init_algorithm=self.init_algorithm,
-                distance=self.distance,
-                n_init=self.n_init,
-                max_iter=self.max_iter,
-                tol=self.tol,
-                verbose=self.verbose,
-                random_state=self._random_state,
-                distance_params=self.distance_params,
-                method="pam",
-            )
-            pam.fit(X[sample_idxs])
-            if pam.inertia_ < best_score:
-                best_pam = pam
+        best_centers = None
+        best_cost = np.inf
 
-        self.labels_ = best_pam.labels_
-        self.inertia_ = best_pam.inertia_
-        self.cluster_centers_ = best_pam.cluster_centers_
-        self.n_iter_ = best_pam.n_iter_
-        self._kmedoids_instance = best_pam
+        for _ in range(self.n_init):
+            centers, cost = self._fit_one_init(X, max_neighbours)
+            if cost < best_cost:
+                best_centers = centers
+                best_cost = cost
 
-    def _score(self, X, y=None):
-        return -self.inertia_
+        labels, inertia = self._assign_clusters(X, best_centers)
+
+        self.labels_ = labels
+        self.inertia_ = inertia
+        self.cluster_centers_ = X[best_centers]
+        self.n_iter_ = 0
 
     @classmethod
     def get_test_params(cls, parameter_set="default"):
@@ -198,7 +190,6 @@ class TimeSeriesCLARA(BaseClusterer):
         parameter_set : str, default="default"
             Name of the set of test parameters to return, for use in tests. If no
             special parameters are defined for a value, will return `"default"` set.
-
 
         Returns
         -------
@@ -212,11 +203,9 @@ class TimeSeriesCLARA(BaseClusterer):
             "n_clusters": 2,
             "init_algorithm": "random",
             "distance": "euclidean",
+            "max_neighbours": None,
             "n_init": 1,
-            "max_iter": 1,
-            "tol": 0.0001,
             "verbose": False,
             "random_state": 1,
-            "n_samples": 10,
-            "n_sampling_iters": 5,
+            "distance_params": None,
         }
