+++ aeon/clustering/tests/test_kernel_k_means.py
@@ -2,7 +2,7 @@
 import numpy as np
 import pytest
 
-from aeon.clustering.kernel_k_means import TimeSeriesKernelKMeans
+from aeon.clustering._kernel_k_means import TimeSeriesKernelKMeans
 from aeon.datasets import load_basic_motions
 from aeon.utils.validation._dependencies import _check_estimator_deps
